// client/redux/auth/auth-stages.tsx
export enum AuthStages {
  INIT = "init",
  SIGNING_IN = "signing-in",
  SIGNED_IN = "signed-in",
  SIGN_IN_FAILURE = "sign-in-failure",
  SIGNING_UP = "signing-up",
  SIGNED_UP = "signed-up",
  SIGN_UP_FAILURE = "sign-up-failure",
  LOGGING_OUT = "logging-out",
  LOGGED_OUT = "logged-out",
  LOGOUT_FAILURE = "logout-failure",
}